@@ -4,17 +4,14 @@
   "description": "Endless Supply of cat pictures (as provided by flickr)",
   "license": "MIT",
   "scripts": {
-    "format": "prettier-eslint \"app/**/*.js\"",
-    "start": "webpack serve --config webpack.dev.js --mode=development",
+    "build": "rollup -c",
+    "watch": "rollup -c -w",
+    "dev": "npm-run-all --parallel start watch",
+    "start": "serve dist",
     "lint": "eslint --ignore-path .gitignore --ext .js,.jsx .",
     "test": "npm run lint; npm run build",
-    "dev": "webpack serve --config webpack.dev.js --mode=development",
-    "prod": "webpack --config webpack.prod.js --mode=production",
-    "build": "webpack --config webpack.dev.js --mode=development",
-    "postversion": "git push --tags && conventional-github-releaser -p angular",
     "build_chrome_package": "rm -rf dist && npm run build && cd dist && zip -r ../infinicatr.zip *",
-    "deploy:surge": "rm -rf dist && npm run build && surge -p dist -d infinicatr.surge.sh",
-    "deploy:apps": "rm -rf dist && npm run build && cd dist && rsync -vazt --delete . rib:/home/apache/vhosts/gavinmogan.com/apps/infinicatr/"
+    "deploy:surge": "rm -rf dist && npm run build && surge -p dist -d infinicatr.surge.sh"
   },
   "repository": {
     "type": "git",
@@ -24,57 +21,44 @@
   "dependencies": {
     "@babel/runtime": "^7.14.6",
     "es6-promise": "^4.2.8",
-    "isomorphic-fetch": "^3.0.0",
     "object-assign": "^4.1.1",
-    "url-search-params": "^1.1.0",
-    "workbox-webpack-plugin": "^6.1.1"
+    "url-search-params": "^1.1.0"
   },
   "devDependencies": {
     "@babel/core": "^7.14.6",
-    "@babel/plugin-proposal-class-properties": "^7.13.0",
-    "@babel/plugin-proposal-decorators": "^7.13.5",
+    "@babel/plugin-proposal-class-properties": "^7.14.5",
+    "@babel/plugin-proposal-decorators": "^7.14.5",
     "@babel/plugin-proposal-object-rest-spread": "^7.14.7",
     "@babel/plugin-transform-runtime": "^7.14.5",
     "@babel/preset-env": "^7.14.7",
+    "@rollup/plugin-commonjs": "^19.0.0",
+    "@rollup/plugin-html": "^0.2.3",
+    "@rollup/plugin-node-resolve": "^11.2.1",
+    "@web/rollup-plugin-html": "^1.8.0",
     "babel-eslint": "^10.1.0",
-    "babel-loader": "^8.2.2",
     "bootstrap-sass": "^3.4.1",
     "cross-env": "^7.0.3",
-    "css-loader": "^5.1.1",
-    "eslint": "^7.21.0",
+    "eslint": "^7.29.0",
     "eslint-config-defaults": "^9.0.0",
-    "eslint-config-standard": "^16.0.2",
-    "eslint-loader": "^4.0.2",
+    "eslint-config-standard": "^16.0.3",
     "eslint-plugin-filenames": "^1.3.2",
-    "eslint-plugin-import": "^2.22.1",
+    "eslint-plugin-import": "^2.23.4",
     "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-promise": "^4.3.1",
+    "eslint-plugin-promise": "^5.1.0",
     "eslint-plugin-standard": "^5.0.0",
     "extend": "^3.0.2",
-    "favicons-webpack-plugin": "^5.0.2",
-    "file-loader": "^6.2.0",
-    "html-loader": "^2.1.2",
-    "html-webpack-plugin": "^5.3.2",
-    "image-webpack-loader": "^7.0.1",
-    "json-loader": "^0.5.7",
-    "loader-utils": "^2.0.0",
     "loaders.css": "^0.1.2",
     "mini-css-extract-plugin": "^2.0.0",
     "mousetrap": "^1.6.5",
-    "mustache-loader": "^1.4.3",
     "node-libs-browser": "^2.2.1",
-    "node-sass": "^5.0.0",
-    "raw-loader": "^4.0.2",
-    "sass-loader": "^11.0.1",
-    "style-loader": "^2.0.0",
-    "surge": "^0.23.0",
-    "url-loader": "^4.1.1",
-    "webpack": "^5.41.1",
-    "webpack-cli": "^4.5.0",
-    "webpack-dev-server": "^3.11.2",
-    "webpack-manifest-plugin": "^3.0.0",
-    "webpack-merge": "^5.7.3",
-    "webpack-pwa-manifest": "^4.3.0",
-    "worker-loader": "^3.0.8"
+    "node-sass": "^6.0.1",
+    "npm-run-all": "^4.1.5",
+    "rollup": "^2.52.6",
+    "rollup-plugin-import-css": "^2.0.1",
+    "rollup-plugin-mustache": "^0.1.0",
+    "rollup-plugin-terser": "^7.0.2",
+    "rollup-plugin-workbox": "^6.1.1",
+    "serve": "^12.0.0",
+    "surge": "^0.23.0"
   }
 }
